evaluator/requirements.txt
# EvalX Python Evaluator Dependencies
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1

# Evaluation Libraries
jsonschema==4.21.1
sentence-transformers==2.5.1

# Database (pgvector)
psycopg2-binary==2.9.9
pgvector==0.2.5

# LLM-as-a-Judge (Groq client)
groq==0.4.2

# Utilities
numpy==1.26.4
